Update Adafruit_ST7796S_kbv.cpp
Added a conditional statement to the "wiring_private.h" header file since some cores of microcontrollers such as the uno rev4 with Wifi does not have such thing. That is why I added a conditional where if a user's controller has it, then they will be able to use it. But if a user does NOT have it, the condition would just ignore the h file.