Update Adafruit_ST7796S_kbv.cpp
Added a conditional where if a user does not have the "wiring_private.h" file then it will be ignored, but if a core of a user does have it, then it will be used. I added this because certain cores of Microcontrollers do not offer this header file such as an Arduino Uno R4 Wifi, And certain Nano controllers as well.